Add Params for AWS (#5)
### Change description ###

In `v2.0.2` I removed the `form` element in favour of using a generic `div` element, as the gem could then be embedded within existing `form` elements. Unfortunately, this broke AWS mode as AWS needs presigned inputs. This version of the gem uses the `Dropzone.options.params`, which effectively adds strings to the upload as if they were inputs to a form.

**Does this PR introduce a breaking change?** (check one with "x")

```
[ ] Yes
[x] No
```